Refactor management model's package-level
