+++ src/test/java/seedu/address/TestApp.java
@@ -9,8 +9,8 @@ import seedu.address.commons.core.Config;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.Lessons;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.modelmanager.management.ManagementModel;
-import seedu.address.model.modelmanager.management.ManagementModelManager;
+import seedu.address.model.modelmanager.ManagementModel;
+import seedu.address.model.modelmanager.ManagementModelManager;
 import seedu.address.model.user.User;
 import seedu.address.storage.UserPrefsStorage;
 import seedu.address.testutil.TestUtil;
