+++ src/test/java/seedu/address/logic/commands/management/ExitCommandTest.java
@@ -7,8 +7,8 @@ import org.junit.Test;
 
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.model.modelmanager.management.ManagementModel;
-import seedu.address.model.modelmanager.management.ManagementModelManager;
+import seedu.address.model.modelmanager.ManagementModel;
+import seedu.address.model.modelmanager.ManagementModelManager;
 
 public class ExitCommandTest {
     private ManagementModel managementModel = new ManagementModelManager();
