+++ src/test/java/seedu/address/model/modelmanager/management/ManagementModelManagerTest.java
@@ -15,6 +15,7 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.Lessons;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.lesson.Lesson;
+import seedu.address.model.modelmanager.ManagementModelManager;
 import seedu.address.model.user.CardSrsData;
 import seedu.address.model.user.User;
