+++ src/test/java/seedu/address/model/modelmanager/management/ManagementModelStub.java
@@ -5,6 +5,7 @@ import java.util.List;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.lesson.Lesson;
+import seedu.address.model.modelmanager.ManagementModel;
 import seedu.address.model.user.CardSrsData;
 import seedu.address.model.user.User;
